requirements.txt
openai
openai[datalib]
faiss-cpu
transformers
torch